Add a variable to update the APT cache.
If ran on a machine with an empty APT cache (on a clean VM or
container), the installation of the packages fails because there is no
metadata to know that the dependencies `apt-transport-https` and
`python-pycurl` are available on the repositories.

For this reason, I've added the `nodejs_packages_update_cache` variable, which
wraps around the `packages_update_cache` variable from the
SimpliField.packages role.